requirements.txt
fastapi
uvicorn
sqlalchemy
openai
streamlit
requests
python-dotenv
